Rename functions to steps and actions to commands
This is the first step in refactoring the CLI
