+++ packages/frontity/src/commands/info.ts
