+++ packages/frontity/src/steps/create-package/steps.ts
@@ -9,7 +9,7 @@ import { fetchPackageVersion } from "../../utils";
 // This function returns true if the directory is not empty and false otherwise.
 export const isDirNotEmpty = ({ packagePath }: Options): Promise<boolean> => {
   return pathExists(join(packagePath));
-}
+};
 
 // This function ensures all directories that needs a package exist.
 export const ensurePackageDir = ({ packagePath }: Options): Promise<void> => {
